Move local modules to local dir
